# Conflicts:
#   frontend/src/assets/main.css

+++ backend/src/main/java/org/ssafy/eeum/domain/family/entity/SupporterId.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @EqualsAndHashCode
 public class SupporterId implements Serializable {
     private Integer user;
-    private Long family;
+    private Integer family;
 }
